requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Web Scraping & Parsing
requests==2.31.0
beautifulsoup4==4.12.3
newspaper3k==0.2.8
lxml==5.1.0

# AI & NLP
openai==1.12.0
anthropic==0.18.1

# Text-to-Speech
elevenlabs==0.2.27
gTTS==2.5.0

# Video Processing
moviepy==1.0.3
Pillow==10.2.0
opencv-python==4.9.0.80

# Audio Processing
pydub==0.25.1

# Subtitle Generation
pysrt==1.1.2

# Utilities
python-dotenv==1.0.1
validators==0.22.0

# Progress & Logging
tqdm==4.66.1
